+++ src/main/g8/be-api/src/gvabe/bootstrap_dao.go
@@ -5,78 +5,87 @@ import (
 	"log"
 	"strings"
 
+	"github.com/btnguyen2k/henge"
 	"github.com/btnguyen2k/prom"
 
 	"main/src/goapi"
-	blogv2 "main/src/gvabe/bov2/blog"
-	userv2 "main/src/gvabe/bov2/user"
-	"main/src/henge"
+	"main/src/gvabe/bov2/blog"
+	"main/src/gvabe/bov2/user"
 	"main/src/utils"
+
+	_ "github.com/jackc/pgx/v4/stdlib"
+	_ "github.com/mattn/go-sqlite3"
 )
 
 func _createSqlConnect(dbtype string) *prom.SqlConnect {
+	timezone := goapi.AppConfig.GetString("timezone")
+	var sqlc *prom.SqlConnect = nil
+	var err error
 	switch dbtype {
 	case "sqlite":
 		dir := goapi.AppConfig.GetString("gvabe.db.sqlite.directory")
 		dbname := goapi.AppConfig.GetString("gvabe.db.sqlite.dbname")
-		return henge.NewSqliteConnection(dir, dbname)
+		sqlc, err = henge.NewSqliteConnection(dir, dbname, timezone, "sqlite3", 10000, nil)
 	case "pg", "pgsql", "postgres", "postgresql":
 		url := goapi.AppConfig.GetString("gvabe.db.pgsql.url")
-		return henge.NewPgsqlConnection(url, goapi.AppConfig.GetString("timezone"))
+		sqlc, err = henge.NewPgsqlConnection(url, timezone, "pgx", 10000, nil)
 	}
-	return nil
+	if err != nil {
+		panic(err)
+	}
+	return sqlc
 }
 
-func _createUserDao(sqlc *prom.SqlConnect) userv2.UserDao {
-	return userv2.NewUserDaoSql(sqlc, userv2.TableUser)
+func _createUserDao(sqlc *prom.SqlConnect) user.UserDao {
+	return user.NewUserDaoSql(sqlc, user.TableUser)
 }
 
-func _createBlogPostDao(sqlc *prom.SqlConnect) blogv2.BlogPostDao {
-	return blogv2.NewBlogPostDaoSql(sqlc, blogv2.TableBlogPost)
+func _createBlogPostDao(sqlc *prom.SqlConnect) blog.BlogPostDao {
+	return blog.NewBlogPostDaoSql(sqlc, blog.TableBlogPost)
 }
 
-func _createBlogCommentDao(sqlc *prom.SqlConnect) blogv2.BlogCommentDao {
-	return blogv2.NewBlogCommentDaoSql(sqlc, blogv2.TableBlogComment)
+func _createBlogCommentDao(sqlc *prom.SqlConnect) blog.BlogCommentDao {
+	return blog.NewBlogCommentDaoSql(sqlc, blog.TableBlogComment)
 }
 
-func _createBlogVoteDao(sqlc *prom.SqlConnect) blogv2.BlogVoteDao {
-	return blogv2.NewBlogVoteDaoSql(sqlc, blogv2.TableBlogVote)
+func _createBlogVoteDao(sqlc *prom.SqlConnect) blog.BlogVoteDao {
+	return blog.NewBlogVoteDaoSql(sqlc, blog.TableBlogVote)
 }
 
 func _createSqlTables(sqlc *prom.SqlConnect, dbtype string) {
 	switch dbtype {
 	case "sqlite":
-		henge.InitSqliteTable(sqlc, userv2.TableUser, map[string]string{userv2.UserCol_MaskUid: "VARCHAR(32)"})
-		henge.InitSqliteTable(sqlc, blogv2.TableBlogPost, map[string]string{
-			blogv2.PostCol_OwnerId: "VARCHAR(32)", blogv2.PostCol_IsPublic: "INT"})
-		henge.InitSqliteTable(sqlc, blogv2.TableBlogComment, map[string]string{
-			blogv2.CommentCol_OwnerId: "VARCHAR(32)", blogv2.CommentCol_PostId: "VARCHAR(32)", blogv2.CommentCol_ParentId: "VARCHAR(32)"})
-		henge.InitSqliteTable(sqlc, blogv2.TableBlogVote, map[string]string{
-			blogv2.VoteCol_OwnerId: "VARCHAR(32)", blogv2.VoteCol_TargetId: "VARCHAR(32)", blogv2.VoteCol_Value: "INT"})
+		henge.InitSqliteTable(sqlc, user.TableUser, map[string]string{user.UserCol_MaskUid: "VARCHAR(32)"})
+		henge.InitSqliteTable(sqlc, blog.TableBlogPost, map[string]string{
+			blog.PostCol_OwnerId: "VARCHAR(32)", blog.PostCol_IsPublic: "INT"})
+		henge.InitSqliteTable(sqlc, blog.TableBlogComment, map[string]string{
+			blog.CommentCol_OwnerId: "VARCHAR(32)", blog.CommentCol_PostId: "VARCHAR(32)", blog.CommentCol_ParentId: "VARCHAR(32)"})
+		henge.InitSqliteTable(sqlc, blog.TableBlogVote, map[string]string{
+			blog.VoteCol_OwnerId: "VARCHAR(32)", blog.VoteCol_TargetId: "VARCHAR(32)", blog.VoteCol_Value: "INT"})
 	case "pg", "pgsql", "postgres", "postgresql":
-		henge.InitPgsqlTable(sqlc, userv2.TableUser, map[string]string{userv2.UserCol_MaskUid: "VARCHAR(32)"})
-		henge.InitPgsqlTable(sqlc, blogv2.TableBlogPost, map[string]string{
-			blogv2.PostCol_OwnerId: "VARCHAR(32)", blogv2.PostCol_IsPublic: "INT"})
-		henge.InitPgsqlTable(sqlc, blogv2.TableBlogComment, map[string]string{
-			blogv2.CommentCol_OwnerId: "VARCHAR(32)", blogv2.CommentCol_PostId: "VARCHAR(32)", blogv2.CommentCol_ParentId: "VARCHAR(32)"})
-		henge.InitPgsqlTable(sqlc, blogv2.TableBlogVote, map[string]string{
-			blogv2.VoteCol_OwnerId: "VARCHAR(32)", blogv2.VoteCol_TargetId: "VARCHAR(32)", blogv2.VoteCol_Value: "INT"})
+		henge.InitPgsqlTable(sqlc, user.TableUser, map[string]string{user.UserCol_MaskUid: "VARCHAR(32)"})
+		henge.InitPgsqlTable(sqlc, blog.TableBlogPost, map[string]string{
+			blog.PostCol_OwnerId: "VARCHAR(32)", blog.PostCol_IsPublic: "INT"})
+		henge.InitPgsqlTable(sqlc, blog.TableBlogComment, map[string]string{
+			blog.CommentCol_OwnerId: "VARCHAR(32)", blog.CommentCol_PostId: "VARCHAR(32)", blog.CommentCol_ParentId: "VARCHAR(32)"})
+		henge.InitPgsqlTable(sqlc, blog.TableBlogVote, map[string]string{
+			blog.VoteCol_OwnerId: "VARCHAR(32)", blog.VoteCol_TargetId: "VARCHAR(32)", blog.VoteCol_Value: "INT"})
 	}
 
 	// user
-	henge.CreateIndex(sqlc, userv2.TableUser, true, []string{userv2.UserCol_MaskUid})
+	henge.CreateIndexSql(sqlc, user.TableUser, true, []string{user.UserCol_MaskUid})
 
 	// blog post
-	henge.CreateIndex(sqlc, blogv2.TableBlogPost, false, []string{blogv2.PostCol_OwnerId})
-	henge.CreateIndex(sqlc, blogv2.TableBlogPost, false, []string{blogv2.PostCol_IsPublic})
+	henge.CreateIndexSql(sqlc, blog.TableBlogPost, false, []string{blog.PostCol_OwnerId})
+	henge.CreateIndexSql(sqlc, blog.TableBlogPost, false, []string{blog.PostCol_IsPublic})
 
 	// blog comment
-	henge.CreateIndex(sqlc, blogv2.TableBlogComment, false, []string{blogv2.CommentCol_OwnerId})
-	henge.CreateIndex(sqlc, blogv2.TableBlogComment, false, []string{blogv2.CommentCol_PostId, blogv2.CommentCol_ParentId})
+	henge.CreateIndexSql(sqlc, blog.TableBlogComment, false, []string{blog.CommentCol_OwnerId})
+	henge.CreateIndexSql(sqlc, blog.TableBlogComment, false, []string{blog.CommentCol_PostId, blog.CommentCol_ParentId})
 
 	// blog vote
-	henge.CreateIndex(sqlc, blogv2.TableBlogVote, true, []string{blogv2.VoteCol_OwnerId, blogv2.VoteCol_TargetId})
-	henge.CreateIndex(sqlc, blogv2.TableBlogVote, false, []string{blogv2.VoteCol_TargetId, blogv2.VoteCol_Value})
+	henge.CreateIndexSql(sqlc, blog.TableBlogVote, true, []string{blog.VoteCol_OwnerId, blog.VoteCol_TargetId})
+	henge.CreateIndexSql(sqlc, blog.TableBlogVote, false, []string{blog.VoteCol_TargetId, blog.VoteCol_Value})
 }
 
 func initDaos() {
@@ -122,7 +131,7 @@ func _initUsers() {
 	}
 	if adminUser == nil {
 		log.Printf("[INFO] Admin user [%s] not found, creating one...", adminUserId)
-		adminUser = userv2.NewUser(goapi.AppVersionNumber, adminUserId, utils.UniqueId())
+		adminUser = user.NewUser(goapi.AppVersionNumber, adminUserId, utils.UniqueId())
 		adminUser.SetPassword(encryptPassword(adminUserId, adminUserPwd)).SetDisplayName(adminUserName).SetAdmin(true)
 		result, err := userDaov2.Create(adminUser)
 		if err != nil {
@@ -163,7 +172,7 @@ Blog content supports <a href="https://en.wikipedia.org/wiki/Markdown" target="_
 
 _Public_ posts are visible to all users for _commenting_ (coming soon) and _voting_.
 `
-		introBlogPost = blogv2.NewBlogPost(goapi.AppVersionNumber, adminUser, true, title, content)
+		introBlogPost = blog.NewBlogPost(goapi.AppVersionNumber, adminUser, true, title, content)
 		introBlogPost.SetId(postId)
 		result, err := blogPostDaov2.Create(introBlogPost)
 		if err != nil {