+++ src/main/g8/be-api/src/gvabe/bov2/user/bo_user_test.go
@@ -7,19 +7,19 @@ import (
 
 func TestNewUser(t *testing.T) {
 	name := "TestNewUser"
-	_appVersion := uint64(1337)
+	_tagVersion := uint64(1337)
 	_id := "admin@local"
 	_maskId := "admin"
 	_pwd := "mypassword"
 	_displayName := "Administrator"
 	_isAdmin := true
-	user := NewUser(_appVersion, _id, _maskId)
+	user := NewUser(_tagVersion, _id, _maskId)
 	if user == nil {
 		t.Fatalf("%s failed: nil", name)
 	}
 	user.SetPassword(_pwd).SetDisplayName(_displayName).SetAdmin(_isAdmin)
-	if appVersion := user.GetAppVersion(); appVersion != _appVersion {
-		t.Fatalf("%s failed: expected app-version to be %#v but received %#v", name, _appVersion, appVersion)
+	if tagVersion := user.GetTagVersion(); tagVersion != _tagVersion {
+		t.Fatalf("%s failed: expected tag-version to be %#v but received %#v", name, _tagVersion, tagVersion)
 	}
 	if id := user.GetId(); id != _id {
 		t.Fatalf("%s failed: expected bo's id to be %#v but received %#v", name, _id, id)
@@ -40,13 +40,13 @@ func TestNewUser(t *testing.T) {
 
 func TestUser_json(t *testing.T) {
 	name := "TestUser_json"
-	_appVersion := uint64(1337)
+	_tagVersion := uint64(1337)
 	_id := "admin@local"
 	_maskId := "admin"
 	_pwd := "mypassword"
 	_displayName := "Administrator"
 	_isAdmin := true
-	user1 := NewUser(_appVersion, _id, _maskId)
+	user1 := NewUser(_tagVersion, _id, _maskId)
 	if user1 == nil {
 		t.Fatalf("%s failed: nil", name)
 	}
@@ -58,8 +58,8 @@ func TestUser_json(t *testing.T) {
 	if err != nil {
 		t.Fatalf("%s failed: %e", name, err)
 	}
-	if user1.GetAppVersion() != user2.GetAppVersion() {
-		t.Fatalf("%s failed: expected %#v but received %#v", name, user1.GetAppVersion(), user2.GetAppVersion())
+	if user1.GetTagVersion() != user2.GetTagVersion() {
+		t.Fatalf("%s failed: expected %#v but received %#v", name, user1.GetTagVersion(), user2.GetTagVersion())
 	}
 	if user1.GetId() != user2.GetId() {
 		t.Fatalf("%s failed: expected %#v but received %#v", name, user1.GetId(), user2.GetId())